requirements.txt
pandas==2.1.3
Pillow==10.1.0
DateTime==5.4
Python-IO==0.3
regex==2023.10.3
